rename gratia_datasource -> gracc_datasource (GRACC-18)
Doing this in a separate PR to make the diff easier to review in #1.
